+++ src/index.ts
@@ -29,5 +29,6 @@ export * from "@/integrations";
 export * from "@/admin";
 export * from "@/auth";
 export * from "@/api";
+export * from "@/asyncapi";
 export * from "@/testing";
 export { VERSION } from "@/version";